requirements.txt
# Core
pydantic>=2.0
pydantic-settings>=2.0

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.28
alembic>=1.12

# HTTP & scraping
playwright>=1.40
langdetect>=1.0

# Observability
langsmith>=0.1.0
langchain-core>=0.1.0

# Gemini (generate module)
google-generativeai>=0.8

# Claude (evaluate module)
anthropic>=0.39

# Logging / utilities
python-dotenv>=1.0

# Testing
pytest>=7.4
pytest-asyncio>=0.21
